Version 0.1
Doing some tests with Spring, JPA and an embedded H2 database.

+++ src/main/java/com/demo/SpringDBwithUI/ProductRepository.java
@@ -7,6 +7,6 @@ public interface ProductRepository extends CrudRepository<Product, Integer> {
 
     List<Product> findByName(String name);
     Product findById(int id);
-    Product findByNameAndWeight(String name, float weight);
+    Product findByNameAndWeight(String name, double weight);
 
 }
